Update CodingTrain Patreon page link
Hey Dan! I'm not sure if I'm right to think that the patreon link is not updated with your own patreon profile. The current link takes me to another profile where I found no sign of you or CodingTrain except the old name `CodingRainbow`. So, I thought that, this might be a mistake because of the migration from `CodingRainbow` to `CodingTrain` OR it's just my mistake thinking like this way.

P.S. Thank you for making me the actual me! Apart from this link fixing, there are a lot more things to talk about how you changed my surrounding and guiding me towards the path I always wanted to walk over and discover. Maybe that's the story for another day :)